Set up PWM support
Right now we are defining the arguments in terms of the frequency
and the duty cycle. On the Edison this is technically the period
and duty cycle, but the Raspberry Pi doesn't actually match the
frequency properly anyway.

These two parameters will need to be standardized in the near
future across Zorg adaptors, so the argument names can line up
and requests to the API can function properly.